Prototype of packer system
nloptr forces all parameters to be packed into a vector.

The packer system automates the generation of type,offset and size
metadata, and provides a nice API to pack / unpack values at the right
position in the packed vector.

The biggest downside is the use of advanced metaprogramming (tuples,
variadic template constructors) and reliance on non obvious "sequencing"
rules of evaluation on make_packer for offset computation.

This is mostly a demonstrator of what can be done using modern C++
features ; it will only be kept if deemed useful enough after discussion.