Investigated error reporting for R to Cpp conversion
Testing the new optimize_full(_weighted)? functions with dummy values can trigger a hard abort of R.
This seems to be due to arma::mat conversion of R "c()" values.
The hard abort cannot be caught and has an unhelpful error message (cannot convert NULL to double).
Going through Rcpp code, hard abort in failed conversion can occur if compiling with NDEBUG defined, which is the case with the default /etc/ configuration on my machine.

I investigated adding some extraction helper functions to ensure returning nice exceptions (and error) on argument mismatch.
However this only partially covers argument conversions, and the useful cases are already covered (list field missing).
As these are internal functions for the package, it is better to document the interface and use them properly at the R level...

Thus the only changes are using const List& to avoid copies when useful.