Nlopt wrapper improvements
Remove template in favor of std::function.
Slightly less efficient, but easier to follow and understand.
-> integrate optimizer construction in the minimize() function.

Add iteration counting in the wrapper itself, returned in the result struct.
This simplifies the computation functions.
The introduction of an OptimData struct is probably simpler to
understand than the lambda cast trick (even if this is functionnally
equivalent).