Fix inclusive arma::span bounds bug
arma::span(start, end) is used to index slices of subvec in packing system.
end is inclusive AND unsigned.
Thus for 0-sized slices of start = 0, end should be -1 which wraps around.

Fixed by using a convoluted way to provide an (offset, size) pair to subvec().

Added some text to explain reasoning.
Another way to explain would be to provide a subvec replacement, but
this would add another template function (not desirable).

Added 0-sized packing/unpacking test to cpp_tests, to detect regressions in the future.