Fix packer compile error.
The compile error came from ambiguities in the complex init syntax rules.
This can fail with old version of the STL which failed to implement the
correct constructor restrictions as stated in the standard.
Fixed by adding unambiguous type specification.

Tested CI with ubuntu 16.04 / 18.04.
It failed due to libc link error with nlopt.so.
This seems to be a binary mismatch (using binaries compiled in a recent
environment in an old one).
If R packages are installed by compilation, this should not happen as it
would be compiled against local versions.
So keeping ubuntu 20.04 to not have this link error.