ZI: outer loop and dimension tests
Add outer loop of optimize_zi.
For now, configuration is shared between outer loop and inner nloptr steps.
Added basic dimensionnality test, also useful as an example of the API.

The test of convergence is complex due to input flexibility.
It has been split as a standalone function to allow testing.

Thorough testing uncovered a bad assumption with named lists :
they are a vector<(value,name)> with insertion order, not a dictionnary.
As the convergence function may be reused elsewhere, it has been
improved to support any order of named lists at the cost of complexity.
The alternative would have been to assume same order and tweak
optimize_zi to get the same ordering of names, but it leads to a brittle
and unclear API.
Runtime cost should be low as named lists are used as struct-like
constructs with fixed (low) size in this context.

step_b solve() problem generates warnings of singular matrix.
A quick R version errors-out for the same reason.